AI Resume Analyzer using RoBERTa – Job Description Matching/requirements.txt
flask==3.0.3
torch==2.3.1
transformers==4.41.2
sentence-transformers==3.0.1
PyPDF2==3.0.1
python-docx==1.1.2
scikit-learn==1.5.0
numpy==1.26.4